client-apps/requirements.txt
llama-stack-client>=0.1.5
python-dotenv
termcolor